Fixed ObjectsScene to remove foreign tables during destruction.
Added the method DatabaseModel::getObject(QString,vector<Object>) to
return the first ocurrency in the lists related to the provided types.

Fixed the View::getObjectCount in order to return 0 when the provided
object type is invalid.

Fixed the ObjectFinderWidget in order to fade out correctly the listed/
not listed elements.