Refactored BaseObject::isCodeDiffersFrom (and its variants in children classes) to use QStringList instead of vector<QString> as parameters
Fixed a crash in diff process caused by ModelExportHelper instance when
canceling the export to dbms due to an exception.

Fixed a crash in SwapObjectsIdsWidget when the user tried to swap
objects using arrow keys.

Allowing swap cluster level object ids in SwapObjectsIdsWidget when the
objects are of the same kind.