Set RW permissions when copying config file with RO.
In package managers such as Nix, application files are stored without write permissions and as a result when copying a default configuration file, the permissions of the file are also kept and as a result pgModeler cannot save the settings.

This commit fixes the following error:
```
Unable to write the file /home/elxreno/.config/pgmodeler/pgmodeler.conf due to one or more errors in the definition generation process!
```

Change-Id: Ib958c59729c4345f8baeab33372391f224ffff5f